+++ packages/eas-cli/src/commands/update/roll-back-to-embedded.ts
@@ -270,7 +270,7 @@ export default class UpdateRollBackToEmbedded extends EasCommand {
               ? [{ label: 'Android update ID', value: newAndroidUpdate.id }]
               : []),
             ...(newIosUpdate ? [{ label: 'iOS update ID', value: newIosUpdate.id }] : []),
-            { label: 'Message', value: updateMessage },
+            { label: 'Message', value: updateMessage ?? '' },
             ...(gitCommitHash
               ? [
                   {
@@ -300,7 +300,7 @@ export default class UpdateRollBackToEmbedded extends EasCommand {
     graphqlClient: ExpoGraphqlClient;
     isGitWorkingTreeDirty: boolean | undefined;
     gitCommitHash: string | undefined;
-    updateMessage: string;
+    updateMessage: string | undefined;
     branchId: string;
     codeSigningInfo: CodeSigningInfo | undefined;
     runtimeVersions: { platform: string; runtimeVersion: string }[];